ExchangeScreen hooks
